+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -100,7 +100,7 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-19T14:53:38.511Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-20T23:11:59.550Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")