+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVG5ZT0YASDCEHKEXMHYEDAP"
+        "01KVKMX4X735T97YRHVZM1FZR3"
     )
     r"""The unique identifier of the data source"""