+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVG5ZT0V939QNZRJTQS3EYS0"
+        "01KVKMX4X2BAHTN028VZ9D07S1"
     )
     r"""The unique identifier of the data source"""