+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatapointop.py
@@ -809,7 +809,7 @@ class RetrieveDatapointEvaluations3(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-06-19T14:54:02.369Z")
+    ] = parse_datetime("2026-06-20T23:12:27.123Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -954,7 +954,7 @@ class RetrieveDatapointEvaluations2(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-06-19T14:54:02.368Z")
+    ] = parse_datetime("2026-06-20T23:12:27.122Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1099,7 +1099,7 @@ class RetrieveDatapointEvaluations1(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-06-19T14:54:02.365Z")
+    ] = parse_datetime("2026-06-20T23:12:27.121Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1212,7 +1212,7 @@ class RetrieveDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-19T14:53:38.511Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-20T23:11:59.550Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")