+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVG5ZT0XE6H1MFK3BADWS2ZX"
+        "01KVKMX4X5Y4V349JQSJ8QTVBQ"
     )
     r"""The unique identifier of the data source"""