+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVG5ZT13QHSHMNF5GTG3A9CY"
+        "01KVKMX4XA5DVFP7Y0DWMMJC7A"
     )
     r"""The unique identifier of the data source"""