+++ packages/orq-rc/RELEASES.md
@@ -8859,3 +8859,13 @@ Based on:
 - [python v4.11.0-rc.35] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.35] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.35 - packages/orq-rc
+
+## 2026-06-20 23:14:25
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.784.0 (2.911.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.36] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.36] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.36 - packages/orq-rc